requirements.txt
streamlit~=1.8.0
setuptools~=60.2.0
pandas~=1.4.1
numpy~=1.22.3
Pillow~=9.0.1
scikit-image~=0.19.2
pydicom~=2.2.2